Cleaned all validation errors. Fixed memory space issues as well as addressed some minor bugs.

+++ Samples/Cube/Shaders/Cube_interop.h
@@ -42,6 +42,6 @@ struct DrawData
 {
     Vertex*   vertices;
     float4x4  mvp;
-    u64       texture;
-    u64       sampler;
+    uint      texture;
+    uint      sampler;
 };
